Admin page built finish (Still some minor kinks need to fix)

+++ src/main/java/com/cs157a1/payMe/Model/CreditCardDao.java
@@ -13,7 +13,7 @@ public interface CreditCardDao {
 	
 	void addcreditCardToDB(CreditCard creditCard, String username);
 	
-	void deletecreditCard(int cardNumber);
+	void deletecreditCard(long cardNumber);
 		
 	void updatecreditCard(CreditCard creditCard);
